# docs/swagger.yaml
basePath: /
definitions:
...
  models.User:
    properties:
      alamat:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      nama:
        type: string
      no_hp:
        type: string
      password:
        type: string
      role:
        type: string
      status:
        description: aktif/nonaktif
        type: string
    type: object
  models.UserInput:
    properties:
      alamat:
        type: string
      email:
        type: string
      nama:
        type: string
      no_hp:
        type: string
      password:
        type: string
      role:
        type: string
      status:
        description: aktif/nonaktif
        type: string
    required:
    - nama
    - email
    - password
    - role
    - status
    type: object
host: backendmbg-production.up.railway.app
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    API documentation untuk Backend GROSIR

    **Sistem Login:**
    - Admin: restu129@gmail.com / restu123
    - Kasir: kasir129@gmail.com / 123456
    - Driver: driver129@gmail.com / 123456

    **Authentication:**
    - Semua endpoint (kecuali login) memerlukan Bearer Token
    - Token didapat dari endpoint /auth/login
    - Format: Authorization: Bearer {token}

    **Role Permissions:**
    - Admin: Akses penuh ke semua fitur
    - Kasir: CRUD produk, pelanggan, pembayaran, lihat riwayat
    - Driver: Lihat pembayaran assigned, update status selesai, cetak surat jalan
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend MBG API
  version: "1.0"
paths:
...
  /users/karyawan:
    get:
      description: Mengambil semua data user/karyawan (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - Karyawan
    post:
      consumes:
      - application/json
      description: Menambah karyawan baru (admin only) - kasir atau driver
      parameters:
      - description: Data karyawan
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserInput'
      produces:
      - application/json
      responses:
        "201":
          description: Karyawan berhasil ditambah
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Request tidak valid
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create new karyawan
      tags:
      - Karyawan
  /users/karyawan/{id}:
    delete:
      description: Hapus karyawan berdasarkan ID (admin only)
      parameters:
      - description: ID karyawan
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Karyawan berhasil dihapus
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete karyawan
      tags:
      - Karyawan
    get:
      description: Mengambil data karyawan berdasarkan ID (admin only)
      parameters:
      - description: ID karyawan
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get karyawan by ID
      tags:
      - Karyawan
    put:
      consumes:
      - application/json
      description: Update data karyawan berdasarkan ID (admin only)
      parameters:
      - description: ID karyawan
        in: path
        name: id
        required: true
        type: string
      - description: Data karyawan yang diupdate
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserInput'
      produces:
      - application/json
      responses:
        "200":
          description: Karyawan berhasil diupdate
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Request tidak valid
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update karyawan
      tags:
      - Karyawan
  /users/karyawan/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update status karyawan menjadi aktif atau nonaktif (admin only)
      parameters:
      - description: ID karyawan
        in: path
        name: id
        required: true
        type: string
      - description: Status karyawan
        in: body
        name: status
        required: true
        schema:
          properties:
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Status karyawan berhasil diupdate
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Request tidak valid
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Karyawan tidak ditemukan
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update karyawan status
      tags:
      - Karyawan
  /users/karyawan/active:
    get:
      description: Mengambil semua data karyawan aktif
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all active karyawan
      tags:
      - Karyawan
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"